Put sudo config in its own file
Current code changes `/etc/sudoers`, which belongs to sudo. If you use this image as a base and somehow have to update sudo, it asks if you want to keep `/etc/sudoers` or overwrite it with the default file. The build then fails cause there is no user input. This patch will change the behavior: it will create a file named `/etc/sudoers.d/neon`, which is included by the `/etc/sudoers` provided by sudo.